pyproject.toml
[project]
name = "one_liner"
version = "0.1.0"

readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.3.5",
    "pyzmq>=27.0.2",
    "ruff>=0.9.9",
]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]

